+++ plone/app/event/vocabularies.py
@@ -1,12 +1,10 @@
 from Products.CMFCore.utils import getToolByName
-from collective.elephantvocabulary import wrap_vocabulary
 from plone.app.event import base
 from plone.app.event import messageFactory as _
 from plone.event.interfaces import IEvent
 from plone.memoize import forever
-from zope.component import getUtility
 from zope.component.hooks import getSite
-from zope.interface import directlyProvides
+from zope.interface import provider
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
@@ -15,76 +13,7 @@ import pytz
 import random
 
 
-def Timezones(context, query=None):
-    """Vocabulary for all timezones.
-    """
-    rpl_keys = base.replacement_zones.keys()
-    tz_list = [SimpleTerm(value=it, title=it)
-               for it in pytz.all_timezones if it not in rpl_keys and (
-                   query is None
-                   or query.lower() in it.lower())]
-    return SimpleVocabulary(tz_list)
-directlyProvides(Timezones, IVocabularyFactory)
-
-
-def AvailableTimezones(context, query=None):
-    """Vocabulary for available timezones, as set by in the controlpanel.
-
-    This vocabulary is based on collective.elephantvocabulary. The reason is,
-    that if timezones are used in events or in user's settings and later
-    retracted by the portal manager, they should still be usable for those
-    objects but not selectable in forms.
-
-    Note: after setting available_timezones, this vocabulary must be
-    reinstantiated to reflect the changes.
-    """
-    # TODO: if the portal_timezone is not in available_timezones, also put it
-    #       in AvailableTimezone vocab.
-    tzvocab = getUtility(IVocabularyFactory,
-                         'plone.app.event.Timezones')(context, query)
-    return wrap_vocabulary(
-        tzvocab,
-        visible_terms_from_registry='plone.app.event.available_timezones'
-    )(context)
-directlyProvides(AvailableTimezones, IVocabularyFactory)
-
-
-def Weekdays(context):
-    """Vocabulary for Weekdays.
-
-    PLEASE NOTE: strftime %w interprets 0 as Sunday unlike the calendar module!
-
-        Note: Context is here a RecordProxy and cannot be used to get the site
-              root. zope.i18n.translate seems not to respect the portal
-              language.
-    """
-
-    # TODO: revisit, use zope.i18n
-    # see: http://weblion.psu.edu/chatlogs/%23plone/2012/08/15.txt
-    # avoid using:
-    # translate = getSite().translate
-    # it breaks tests. it's defined in:
-    # Products.CMFPlone.skins.plone_scripts.translate.py
-    # better use:
-    # from zope.i18n import translate
-    translate = getToolByName(getSite(), 'translation_service').translate
-
-    domain = 'plonelocales'
-    items = [
-        (translate(u'weekday_mon', domain=domain, default=u'Monday'), 0),
-        (translate(u'weekday_tue', domain=domain, default=u'Tuesday'), 1),
-        (translate(u'weekday_wed', domain=domain, default=u'Wednesday'), 2),
-        (translate(u'weekday_thu', domain=domain, default=u'Thursday'), 3),
-        (translate(u'weekday_fri', domain=domain, default=u'Friday'), 4),
-        (translate(u'weekday_sat', domain=domain, default=u'Saturday'), 5),
-        (translate(u'weekday_sun', domain=domain, default=u'Sunday'), 6),
-    ]
-
-    items = [SimpleTerm(i[1], i[1], i[0]) for i in items]
-    return SimpleVocabulary(items)
-directlyProvides(Weekdays, IVocabularyFactory)
-
-
+@provider(IVocabularyFactory)
 @forever.memoize
 def EventTypes(context):
     # TODO: refactor for DX
@@ -122,9 +51,9 @@ def EventTypes(context):
         tmp_folder.__parent__.manage_delObjects([tmp_folder_id])
 
     return SimpleVocabulary.fromValues(event_types)
-directlyProvides(EventTypes, IVocabularyFactory)
 
 
+@provider(IVocabularyFactory)
 def SynchronizationStrategies(context):
     """Vocabulary for icalendar synchronization strategies.
 
@@ -145,4 +74,3 @@ def SynchronizationStrategies(context):
     ]
     items = [SimpleTerm(title=i[0], value=i[1]) for i in items]
     return SimpleVocabulary(items)
-directlyProvides(SynchronizationStrategies, IVocabularyFactory)